+++ packages/ml/src/FEGrid/ml_pde.c
@@ -64,16 +64,16 @@ int ML_PDE_GenMat(MLI_Solver *solver, int N_nodes)
   
    nbytes = nnode_part_1d * 3 * sizeof(int*);
 #ifdef ML_CPP
-   ML_memory_alloc((void**) &square, nbytes, "AP1");
+   ML_memory_alloc((void**) &square, (unsigned int) nbytes, "AP1");
 #else
-   ML_memory_alloc((void*) &square, nbytes, "AP1");
+   ML_memory_alloc((void*) &square, (unsigned int) nbytes, "AP1");
 #endif
    nbytes = nnode_part_1d * 3 * sizeof(int);
    for ( i = 0; i < nnode_part_1d*3; i++ )
 #ifdef ML_CPP
-      ML_memory_alloc((void**) &(square[i]), nbytes, "AP2");
+     ML_memory_alloc((void**) &(square[i]), (unsigned int) nbytes, "AP2");
 #else
-      ML_memory_alloc((void*) &(square[i]), nbytes, "AP2");
+     ML_memory_alloc((void*) &(square[i]), (unsigned int) nbytes, "AP2");
 #endif
 
    for ( j = 0; j < 3; j++ )
